Moved buffer builder classes into data dir
