+++ js/data/symbol_buffer_builder.js
@@ -3,17 +3,17 @@
 var Point = require('point-geometry');
 
 var BufferBuilder = require('./buffer_builder');
-var ElementGroups = require('../element_groups');
-var Anchor = require('../../symbol/anchor');
-var getAnchors = require('../../symbol/get_anchors');
-var resolveTokens = require('../../util/token');
-var Quads = require('../../symbol/quads');
-var Shaping = require('../../symbol/shaping');
-var resolveText = require('../../symbol/resolve_text');
-var mergeLines = require('../../symbol/mergelines');
-var clipLine = require('../../symbol/clip_line');
-var CollisionFeature = require('../../symbol/collision_feature');
-var util = require('../../util/util');
+var ElementGroups = require('./element_groups');
+var Anchor = require('../symbol/anchor');
+var getAnchors = require('../symbol/get_anchors');
+var resolveTokens = require('../util/token');
+var Quads = require('../symbol/quads');
+var Shaping = require('../symbol/shaping');
+var resolveText = require('../symbol/resolve_text');
+var mergeLines = require('../symbol/mergelines');
+var clipLine = require('../symbol/clip_line');
+var CollisionFeature = require('../symbol/collision_feature');
+var util = require('../util/util');
 
 var shapeText = Shaping.shapeText;
 var shapeIcon = Shaping.shapeIcon;
